+++ src/image_io.cpp
@@ -78,7 +78,7 @@ size_t Image::size() const {
 /// @param y y-coordinate
 /// @param c channel
 /// @return Pixel channel value
-float Image::get_pixel(int x, int y, Channel c) const {
+double Image::get_pixel(int x, int y, Channel c) const {
     return data[(y * width + x) * channels + c];
 }
 
@@ -87,7 +87,7 @@ float Image::get_pixel(int x, int y, Channel c) const {
 /// @param y y-coordinate
 /// @param c channel
 /// @param value Pixel channel value
-void Image::set_pixel(int x, int y, Channel c, float value) {
+void Image::set_pixel(int x, int y, Channel c, double value) {
     data[(y * width + x) * channels + c] = value;
 }
 
@@ -100,8 +100,8 @@ bool Image::save(const char* filename, const ImageFormat format) const {
     std::vector<uint8_t> data8(this->size());
     for (size_t i = 0; i < this->size(); ++i) {
         // Clamp the pixel values to [0, 255]
-        uint8_t result = static_cast<uint8_t>(
-            std::max(0.0f, std::min(255.0f, this->data[i])));
+        uint8_t result =
+            static_cast<uint8_t>(std::max(0.0, std::min(255.0, this->data[i])));
         data8[i] = result;
     }
 
@@ -136,6 +136,6 @@ bool Image::save(const std::string filename, const ImageFormat format) const {
 /// @param y y-coordinate
 /// @param c channel
 /// @return Pixel channel value
-float Image::operator()(int x, int y, Channel c) const {
+double Image::operator()(int x, int y, Channel c) const {
     return get_pixel(x, y, c);
 }
